add MIX example and refactor mix_sample_strategy
